refactor: moved admin component to admin domain
