+++ src/app/admin/categories/page.tsx
@@ -3,9 +3,8 @@
 import { useEffect, useState } from "react";
 
 import { TGetAllCategories, getAllCategories } from "@/actions/category/category";
-import AddCategoryGroup from "@/components/admin/category/addCategoryGroup";
-import CatGroupRow from "@/components/admin/category/rowGroup";
-
+import AddCategoryGroup from "@/domains/admin/components/category/addCategoryGroup";
+import CatGroupRow from "@/domains/admin/components/category/rowGroup";
 
 const AdminCategories = () => {
   const [allCategories, setAllCategories] = useState<TGetAllCategories[]>([]);
