+++ src/app/admin/layout.tsx
@@ -2,7 +2,7 @@ import { Metadata } from "next";
 import { redirect } from "next/navigation";
 import { getServerSession } from "next-auth";
 
-import AdminSidebar from "@/components/admin/sideBar";
+import AdminSidebar from "@/domains/admin/components/sideBar";
 import { authOptions } from "@/shared/lib/authOptions";
 
 export const metadata: Metadata = {
